🐛 fix(client): fix warning code and divide recoil code
전역으로 관리하는 변수를 따로 빼주고, 쌓여있는 warning코드를 삭제하였습니다.
