+++ packages/client/src/molecules/ChatSection/InfoBoxRoomFunction.tsx
@@ -1,21 +1,15 @@
 import { styled } from '@mui/material/styles';
-import CustomIconButton from 'client/src/atoms/button/icon/CustomIconButtion';
 import BlockIcon from '@mui/icons-material/Block';
 import SportsKabaddiIcon from '@mui/icons-material/SportsKabaddi';
 import PersonIcon from '@mui/icons-material/Person';
 import MeetingRoomIcon from '@mui/icons-material/MeetingRoom';
-import PasswordInput from 'client/src/atoms/input/passwordBox';
-import {
-  CustomIconProps,
-  LinkIconProps,
-  LinkIconResource,
-} from 'client/src/types/Link.type';
-import LinkPageIconButton from 'client/src/atoms/button/linkPage/LinkPageIconButton';
-import { CHANNELURL, SERVERURL } from 'client/src/configs/Link.url';
 import axios from 'axios';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { userDataAtom } from '../../pages/PingpongRoutePage';
+import { CHANNELURL, SERVERURL } from '../../configs/Link.url';
+import { CustomIconProps } from '../../types/Link.type';
+import CustomIconButton from '../../atoms/button/icon/CustomIconButtion';
 
 const InfoBoxFunctionLayout = styled('div')(({ theme }) => ({
   width: '30%',
@@ -26,7 +20,6 @@ const InfoBoxFunctionLayout = styled('div')(({ theme }) => ({
 }));
 //향후 상태관리를 추가하여 조건에 따라 아이콘을 보이게 또는 안보이게 처리해줄 것 입니다.
 function InfoBoxRoomFunctionModule() {
-  //[수정사항] 동환님이 유저 작업끝내면 바꿀 것
   const userData = useRecoilValue(userDataAtom);
   const navigate = useNavigate();
   const { roomId } = useParams();
