+++ packages/client/src/molecules/ChatSection/RoomInfoBoxPassword.tsx
@@ -1,31 +1,17 @@
 import { styled } from '@mui/material/styles';
 import CustomIconButton from '../../atoms/button/icon/CustomIconButtion';
-import {
-  Avatar,
-  Typography,
-  IconButton,
-  Button,
-  TextField,
-  Input,
-  InputLabel,
-  FormControl,
-} from '@mui/material';
-import InsertPhotoIcon from '@mui/icons-material/InsertPhoto';
 import PasswordInput from '../../atoms/input/passwordBox';
 import LockIcon from '@mui/icons-material/Lock';
 import LockOpenIcon from '@mui/icons-material/LockOpen';
 import { CustomIconProps } from '../../types/Link.type';
 import { DM, PROTECTED } from '../../configs/RoomType';
-import axios from 'axios';
-import { SERVERURL } from '../../configs/Link.url';
 import { useEffect, useState } from 'react';
 import { RoomInfoSet } from '../../types/Room.type';
 import { ChangeRoomInfo } from '../../organisms/ChatMainSection/EnteredChatRoomInfo';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { userDataAtom } from '../../pages/PingpongRoutePage';
 import { OWNER } from '../../configs/userType';
-import { userAuth } from '../../template/ChatMainSection/EnteredChatRoomTemplate';
+import { userAuth } from '../../recoil/chat.recoil';
 
 const InfoBoxPasswordLayout = styled('div')(({ theme }) => ({
   width: '30%',
@@ -41,9 +27,8 @@ const PasswordIconLayout = styled('div')(({ theme }) => ({
 
 function InfoBoxPasswordModule(props: { roomInfoSet: RoomInfoSet }) {
   const roomInfoSet = props.roomInfoSet;
-  const { roomInfo, handler } = roomInfoSet;
-  const userData = useRecoilValue(userDataAtom);
-  const { name, type, image } = roomInfo;
+  const { roomInfo } = roomInfoSet;
+  const { type } = roomInfo;
   const { roomId } = useParams();
   const [password, setPassword] = useState('');
   const [changePassword, setChangePassword] = useState<boolean>(false);
