+++ packages/client/src/molecules/ProfileSection/UserChatParticipants.tsx
@@ -14,9 +14,9 @@ import LinkPageIconButton from '../../atoms/button/linkPage/LinkPageIconButton';
 import { UserProfileBoxDataType } from '../../types/Profile.type';
 import { useRecoilValue } from 'recoil';
 import { PROFILEURL } from '../../configs/Link.url';
-import { userAuth } from '../../template/ChatMainSection/EnteredChatRoomTemplate';
 import { ADMIN, OWNER } from '../../configs/userType';
 import { userDataAtom } from '../../pages/PingpongRoutePage';
+import { userAuth } from '../../recoil/chat.recoil';
 
 const UserProfileLayout = styled(Badge)(({ theme }) => ({
   marginLeft: '4%',
@@ -47,7 +47,7 @@ const customProps: CustomIconProps = {
 //향후 상태관리를 추가하여 조건에 따라 아이콘을 보이게 또는 안보이게 처리해줄 것 입니다.
 function UserChatParticipantsBox(props: { participantInfo: any }) {
   const participantInfo = props.participantInfo;
-  const { user, auth, status, blocked } = participantInfo;
+  const { user, auth } = participantInfo;
   const userType = useRecoilValue(userAuth);
   const userData = useRecoilValue(userDataAtom);
   const userInfo: UserProfileBoxDataType = {
