+++ packages/client/src/organisms/ChatSidebar/JoinedChatRoomList.tsx
@@ -1,20 +1,16 @@
 import { styled } from '@mui/material/styles';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { atom, useRecoilState, useRecoilValue } from 'recoil';
-import ListGenerate from '../../atoms/list/ListGenerate';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import {
   ListGenerateLayout,
   ListLayout,
   ListUlLayout,
 } from '../../atoms/list/styles/ListStylesCSS';
-import { SERVERURL } from '../../configs/Link.url';
-import ProfileChatRoomBox from '../../molecules/ProfileSection/ProfileChatRoomBox';
 import UserProfileBox from '../../molecules/ProfileSection/UserProfileBox';
-import { chatRoomList, getJoinedChatList } from '../../pages/ChatChannelPage';
 import { userDataAtom } from '../../pages/PingpongRoutePage';
+import { chatRoomList, getJoinedChatList } from '../../recoil/chat.recoil';
 import { UserProfileBoxDataType } from '../../types/Profile.type';
-import { RoomList } from '../../types/Room.type';
 
 const JoinedChatRoomLayout = styled('div')(({ theme }) => ({
   height: '35.3%',
@@ -41,14 +37,14 @@ const BoxLayout = styled('div')(({ theme }) => ({
 
 function JoinedChatRoomModule() {
   const userData = useRecoilValue(userDataAtom);
-  const [Roomlist, setRoomList] = useRecoilState(chatRoomList);
+  const [roomlist, setRoomList] = useRecoilState(chatRoomList);
   const joinedChatList = useRecoilValue(getJoinedChatList(userData.id));
   const navigate = useNavigate();
   useEffect(() => {
     setRoomList(joinedChatList.channelList);
   }, [joinedChatList.channelList, setRoomList]);
 
-  const listElement: React.ReactElement[] = Roomlist.map((room: any) => {
+  const listElement: React.ReactElement[] = roomlist.map((room: any) => {
     const profileData: UserProfileBoxDataType = {
       nickname: room.name,
       image: room.image,
